Rename botEngine module
