+++ botCommands/giphy.js
@@ -1,7 +1,3 @@
-// const { giphy: { apiKey }} = require("../config.js");
-// const giphy = require("giphy-api")(apiKey);
-// const { registerBotCommand } = require("../bot-engine.js");
-
 async function chooseRandomGif(searchTerm) {
   return new Promise(async (resolve, reject) => {
     try {
